+++ spec/support/register-css-inline-loader.mjs
@@ -1,0 +1,3 @@
+import { register } from 'node:module'
+
+register(new URL('./css-inline-loader.mjs', import.meta.url))